+++ runTests.js
@@ -1,5 +1,5 @@
-require('babel-core/register');
-require('babel-polyfill');
+require('@babel/register');
+require('@babel/polyfill');
 
 const MiaJs = require('mia-js-core').Run;
 const Jasmine = require('jasmine');